requirements.txt
asgiref==3.8.1
Django==5.1.5
django-appconf==1.0.6
django-imagekit==5.0.0
djangorestframework==3.15.2
pilkit==3.0
pillow==11.1.0
python-dotenv==1.0.1
python-slugify==8.0.4
slugify==0.0.1
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.2
